Revert "Swing"
