--- .idea/artifacts/22_07_06_Caged_T5_Capstone_jar.xml
@@ -1,14 +1,0 @@
-<component name="ArtifactManager">
-  <artifact type="jar" name="22.07.06-Caged-T5-Capstone:jar">
-    <output-path>$PROJECT_DIR$/out/artifacts/22_07_06_Caged_T5_Capstone_jar</output-path>
-    <root id="archive" name="22.07.06-Caged-T5-Capstone.jar">
-      <element id="module-output" name="22.07.06-Caged-T5-Capstone" />
-      <element id="extracted-dir" path="$MAVEN_REPOSITORY$/com/fasterxml/jackson/dataformat/jackson-dataformat-yaml/2.13.0/jackson-dataformat-yaml-2.13.0.jar" path-in-jar="/" />
-      <element id="extracted-dir" path="$MAVEN_REPOSITORY$/com/fasterxml/jackson/core/jackson-databind/2.13.0/jackson-databind-2.13.0.jar" path-in-jar="/" />
-      <element id="extracted-dir" path="$MAVEN_REPOSITORY$/com/fasterxml/jackson/core/jackson-annotations/2.13.0/jackson-annotations-2.13.0.jar" path-in-jar="/" />
-      <element id="extracted-dir" path="$MAVEN_REPOSITORY$/org/yaml/snakeyaml/1.28/snakeyaml-1.28.jar" path-in-jar="/" />
-      <element id="extracted-dir" path="$MAVEN_REPOSITORY$/com/fasterxml/jackson/core/jackson-core/2.13.0/jackson-core-2.13.0.jar" path-in-jar="/" />
-      <element id="extracted-dir" path="$MAVEN_REPOSITORY$/com/fasterxml/jackson/datatype/jackson-datatype-jsr310/2.13.0/jackson-datatype-jsr310-2.13.0.jar" path-in-jar="/" />
-    </root>
-  </artifact>
-</component>
