--- .idea/libraries/fasterxml_jackson_core_databind.xml
@@ -1,12 +1,0 @@
-<component name="libraryTable">
-  <library name="fasterxml.jackson.core.databind" type="repository">
-    <properties maven-id="com.fasterxml.jackson.core:jackson-databind:2.13.0" />
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/fasterxml/jackson/core/jackson-databind/2.13.0/jackson-databind-2.13.0.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/com/fasterxml/jackson/core/jackson-annotations/2.13.0/jackson-annotations-2.13.0.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/com/fasterxml/jackson/core/jackson-core/2.13.0/jackson-core-2.13.0.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
