+++ src/main/java/com/caged/Character.java
