+++ src/main/java/com/caged/CharacterNPCGetter.java
