+++ src/main/java/com/caged/CharacterPlayer.java
