+++ src/main/java/com/caged/DictionaryGetter.java
