+++ src/main/java/com/caged/GameControl.java
@@ -82,9 +82,9 @@ public class GameControl<K, V> {
             // This will become a JPanels w/ text field output (remove println portions)
             System.out.println("\nThings seen in room: ");
             //How is it printing
-//            KeyValueParser.key(node.get("room").get(playerLocation).get("Inventory"));
+            KeyValueParser.key(node.get("room").get(playerLocation).get("Inventory"));
             System.out.println("\nPeople seen in room: ");
-//            KeyValueParser.key(node.get("room").get(playerLocation).get("NPCs"));
+            KeyValueParser.key(node.get("room").get(playerLocation).get("NPCs"));
             System.out.println("\nDirections you can move: ");
             // TODO: provides movable directions to screen:
             KeyValueParser.locationKeyValue(node.get("room").get(playerLocation).get("Moves"), player, doors);
