+++ src/main/java/com/caged/InventoryGlobal.java
@@ -8,7 +8,6 @@ public class InventoryGlobal {
    public static List<String> roomInvList = new ArrayList<>();
    public static List<String> npcList = new ArrayList<>();
    public static List<String> locationList = new ArrayList<>();
-   public static List <String> itemsList = new ArrayList<>();
 
 
 }
