+++ src/main/java/com/caged/PlayerStatus.java
